Fixed View Models to handle No Internet Connection

+++ app/src/main/java/com/example/sampleappkd/viewmodel/DeleteDoctorViewModel.kt
@@ -4,9 +4,11 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.sampleappkd.model.DeleteDoctorResponse
+import com.example.sampleappkd.network.NetworkService
 import com.example.sampleappkd.repository.DeleteDoctorRepository
 import com.example.sampleappkd.util.Resource
 import kotlinx.coroutines.launch
+import okio.IOException
 import retrofit2.Response
 
 class DeleteDoctorViewModel (val deleteDoctorRepository: DeleteDoctorRepository) : ViewModel() {
@@ -14,9 +16,24 @@ class DeleteDoctorViewModel (val deleteDoctorRepository: DeleteDoctorRepository)
     val data: MutableLiveData<Resource<DeleteDoctorResponse>> = MutableLiveData()
 
     fun deleteDoctor(id: String?) = viewModelScope.launch {
+        deleteDoctorCall(id)
+    }
+
+    private suspend fun deleteDoctorCall(id: String?) {
         data.postValue(Resource.Loading())
-        val response = deleteDoctorRepository.deleteDoctor(id)
-        data.postValue(handleDeleteDoctorResponse(response))
+        try {
+            if(NetworkService.instance.isOnline()) {
+                val response = deleteDoctorRepository.deleteDoctor(id)
+                data.postValue(handleDeleteDoctorResponse(response))
+            } else {
+                data.postValue(Resource.Error("No Internet Connection"))
+            }
+        } catch(t: Throwable) {
+            when(t) {
+                is IOException -> data.postValue(Resource.Error("Network Failure"))
+                else -> data.postValue(Resource.Error("An Error Occurred"))
+            }
+        }
     }
 
     private fun handleDeleteDoctorResponse(response: Response<DeleteDoctorResponse>) : Resource<DeleteDoctorResponse> {
